common: add fopen() wrapper
Move the fopen() wrapper from slidetool-util into common for future use.
Add the manual FD_CLOEXEC setting code from OpenSlide.

Expand the pre-commit fopen() prohibition to the whole codebase.  Allow
the prohibited function check to be overridden the same way as the wrapped
function check so we can implement the wrapper.

make-tables still can't use the wrapper since it's compiled for the build
machine and glib is unavailable.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>